+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Fixed a bug where icons within secondary buttons were illegible when active.
 
 ## Unreleased (4.5)
+- Added `craft\events\AssetBundleEvent`.
+- Added `craft\web\View::EVENT_AFTER_REGISTER_ASSET_BUNDLE`.
 - Fixed multiple issues with Selectize inputs.
 
 ## 4.5.0-beta.1 - 2023-08-10